requirements.txt
fastapi==0.111.0
uvicorn==0.29.0

pdfplumber==0.11.0
python-docx==1.1.0
requests

sentence-transformers==2.7.0
faiss-cpu==1.7.4
transformers==4.41.1
torch  # For transformers to work
